+++ src/form-credit-card-number-field.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {useForm, useSyncFieldValue} from './form-context';
+import {useForm, useFieldProps} from './form-context';
 import {useTheme} from './hooks';
 import {
     getCreditCardNumberLength,
@@ -170,17 +170,7 @@ export const FormCreditCardNumberField: React.FC<FormCreditCardNumberFieldProps>
     ...rest
 }) => {
     const {texts} = useTheme();
-    const {
-        jumpToNext,
-        rawValues,
-        setRawValue,
-        values,
-        setValue,
-        formStatus,
-        formErrors,
-        setFormError,
-        register,
-    } = useForm();
+    const {jumpToNext, rawValues, values, setFormError} = useForm();
 
     const validate = (value: string | undefined, rawValue: string) => {
         const error = texts.formCreditCardNumberError;
@@ -207,26 +197,30 @@ export const FormCreditCardNumberField: React.FC<FormCreditCardNumberFieldProps>
 
     const processValue = (s: string) => s.replace(/\s/g, '');
 
-    useSyncFieldValue({name, value, defaultValue, processValue});
+    const fieldProps = useFieldProps({
+        name,
+        value,
+        defaultValue,
+        processValue,
+        helperText,
+        optional,
+        error,
+        disabled,
+        onBlur,
+        validate,
+        onChange,
+        onChangeValue,
+    });
 
     return (
         <TextFieldBase
             {...rest}
-            inputRef={(field) => register({name, field, validate})}
-            disabled={disabled || formStatus === 'sending'}
-            error={error || !!formErrors[name]}
-            helperText={formErrors[name] ?? helperText}
-            name={name}
-            required={!optional}
-            value={value ?? rawValues[name] ?? ''}
+            {...fieldProps}
             maxLength={getCreditCardNumberLength(values[name]) + 3} // We have to take in account formatting spaces
             onChange={(event) => {
+                fieldProps.onChange(event);
                 const rawValue = event.currentTarget.value;
                 const value = processValue(rawValue);
-                setRawValue({name, value: rawValue});
-                setValue?.({name, value});
-                onChange?.(event);
-                onChangeValue?.(value, rawValue);
                 if (value.length >= getCreditCardNumberLength(value)) {
                     const error = validate?.(value, rawValue);
                     if (error) {
@@ -234,13 +228,7 @@ export const FormCreditCardNumberField: React.FC<FormCreditCardNumberFieldProps>
                     } else {
                         jumpToNext(name);
                     }
-                } else {
-                    setFormError({name, error: ''});
                 }
-            }}
-            onBlur={(e) => {
-                setFormError({name, error: validate?.(values[name], rawValues[name])});
-                onBlur?.(e);
             }}
             inputComponent={CreditCardInput}
             autoComplete={autoComplete}