+++ src/form-integer-field.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {useForm, useSyncFieldValue} from './form-context';
+import {useFieldProps} from './form-context';
 import {useTheme} from './hooks';
 
 import type {CommonFormFieldProps} from './form';
@@ -45,16 +45,6 @@ export const FormIntegerField: React.FC<FormIntegerFieldProps> = ({
     ...rest
 }) => {
     const {texts} = useTheme();
-    const {
-        rawValues,
-        setRawValue,
-        values,
-        setValue,
-        formStatus,
-        formErrors,
-        setFormError,
-        register,
-    } = useForm();
 
     const validate = (value: string | undefined, rawValue: string) => {
         if (!value) {
@@ -65,32 +55,20 @@ export const FormIntegerField: React.FC<FormIntegerFieldProps> = ({
 
     const processValue = (value: string) => value.trim();
 
-    useSyncFieldValue({name, value, defaultValue, processValue});
+    const fieldProps = useFieldProps({
+        name,
+        value,
+        defaultValue,
+        processValue,
+        helperText,
+        optional,
+        error,
+        disabled,
+        onBlur,
+        validate,
+        onChange,
+        onChangeValue,
+    });
 
-    return (
-        <TextFieldBase
-            {...rest}
-            inputRef={(field) => register({name, field, validate})}
-            disabled={disabled || formStatus === 'sending'}
-            error={error || !!formErrors[name]}
-            helperText={formErrors[name] || helperText}
-            name={name}
-            required={!optional}
-            value={value ?? rawValues[name] ?? ''}
-            onChange={(event) => {
-                const rawValue = event.currentTarget.value;
-                const value = processValue(rawValue);
-                setRawValue({name, value: rawValue});
-                setValue({name, value});
-                setFormError({name, error: ''});
-                onChange?.(event);
-                onChangeValue?.(value, rawValue);
-            }}
-            onBlur={(e) => {
-                setFormError({name, error: validate?.(values[name], rawValues[name])});
-                onBlur?.(e);
-            }}
-            inputComponent={IntegerInput}
-        />
-    );
+    return <TextFieldBase {...rest} {...fieldProps} inputComponent={IntegerInput} />;
 };